requirements.txt
streamlit
requests
beautifulsoup4
pandas
openai
python-dotenv
